moved generation scripts to scripts/default.py, other stuff too
i still need to think about the way this is gonna work in the end. i'm
probably going to make jade.py read and run the script files as a script
instead of importing, so that they can be kept as external and editable
things for non-python users. who even knows